agrego y corrigo mis controladores, asi tambien mis modelos y el validator le agrego lo q me faltaba. agrego los helpers correspondientes y aplico la logica del auth.controllers y sus rutas (falta middleware)

+++ src/controllers/user.controllers.js
@@ -1,29 +1,12 @@
 import userModel from "../models/user.models.js";
 
-const createUser = async (req, res) =>{
-    const {username, email, password, role} = req.body;
-    try {
-        const newUser = await userModel.create({
-            username,
-            email,
-            password,
-            role
-        })
-        res.status(201).json({
-            msg: "Usuario creado correctamente",
-            data: newUser
-        })
-    } catch (error) {
-        res.status(500).json({msg: "Error al crear usuario"})
-    }
-}
-
 const getAllUsers = async (req, res) => {
     try {
         const users = await userModel.find();
         res.status(200).json({
             msg: "Lista de usuarios",
-            data: users
+            data: users,
+            ok: true
         });
     } catch (error) {
         res.status(500).json({ msg: "Error al obtener usuarios", error });
@@ -31,25 +14,40 @@ const getAllUsers = async (req, res) => {
 }
 
 const getUserByID = async (req, res) => {
+    const existingUser = await userModel.findById(req.params.id);
+    if (!existingUser) {
+        return res.status(404).json({msg: "El usuario no existe"});
+    }
     try {
-        const user = await userModel.findById(req.params.id);
+        const user = await userModel.findById(req.params.id).populate('articles').populate('comments');
         res.status(200).json({
             msg: "Usuario correspondinte:",
-            data: user
+            data: user,
+            ok: true
         })
     } catch (error) {
-        res.status(500).json({msg: "Error al obtener usuario", error})
+        res.status(500).json({
+            msg: "Error al obtener usuario",
+            error,
+            ok: false
+        })
     }
 }
 
 const updateUser = async (req, res) => {
     try {
-        const user = await userModel.findByIdAndUpdate(
+        const existingUser = await userModel.findById(req.params.id);
+        if (!existingUser) {
+            return res.status(404).json({msg: "El usuario no existe"});
+        }
+        const updatedUser = await userModel.findByIdAndUpdate(
             req.params.id, 
-            req.body)
+            req.body
+        );
         res.status(200).json({
             msg: "usuario actualizado correctamente",
-            data: user
+            data: updatedUser,
+            ok: true
         })
     } catch (error) {
         res.status(500).json({msg: "Error al actualizar el usuario"})
@@ -58,10 +56,15 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
     try {
+        const existingUser = await userModel.findById(req.params.id);
+        if (!existingUser) {
+            return res.status(404).json({msg: "El usuario no existe"});
+        }
         const user = await userModel.findByIdAndDelete(req.params.id);
         res.status(200).json({
             msg: "Usuario eliminado correctamente",
-            data: user
+            data: user,
+            ok: true
         });
     } catch (error) {
         res.status(500).json({msg: "Error al eliminar usuario"})
